+++ templates/product/access-plan-button.php
@@ -7,6 +7,7 @@
  * @since     3.23.0
  * @version   3.23.0
  */
+defined( 'ABSPATH' ) || exit;
 ?>
 <?php if ( get_current_user_id() && $plan->has_free_checkout() && $plan->is_available_to_user() ) : ?>
 	<?php llms_get_template( 'product/free-enroll-form.php', compact( 'plan' ) ); ?>